feat(Remove Bootstrap): Remove Offcanvas bootstrap code
