+++ packages/bootstrap-vue-3/src/BootstrapVue.ts
@@ -79,7 +79,7 @@ import BNavbar from './components/BNavbar/BNavbar.vue'
 import BNavbarBrand from './components/BNavbar/BNavbarBrand.vue'
 import BNavbarNav from './components/BNavbar/BNavbarNav.vue'
 import BNavbarToggle from './components/BNavbar/BNavbarToggle.vue'
-import BOffcanvas from './components/BOffcanvas.vue'
+import BOffcanvas from './components/BOffcanvas/BOffcanvas.vue'
 import BOverlay from './components/BOverlay/BOverlay.vue'
 import BPagination from './components/BPagination/BPagination.vue'
 import BPlaceholder from './components/BPlaceholder/BPlaceholder.vue'
