+++ packages/bootstrap-vue-3/src/components/index.ts
@@ -69,7 +69,7 @@ import BNavbar from './BNavbar/BNavbar.vue'
 import BNavbarBrand from './BNavbar/BNavbarBrand.vue'
 import BNavbarNav from './BNavbar/BNavbarNav.vue'
 import BNavbarToggle from './BNavbar/BNavbarToggle.vue'
-import BOffcanvas from './BOffcanvas.vue'
+import BOffcanvas from './BOffcanvas/BOffcanvas.vue'
 import BOverlay from './BOverlay/BOverlay.vue'
 import BPagination from './BPagination/BPagination.vue'
 import BPlaceholder from './BPlaceholder/BPlaceholder.vue'
