+++ packages/bootstrap-vue-3/src/styles/styles.scss
@@ -13,6 +13,7 @@
 @import "../components/BToast/index";
 @import "../components/BFormInput/form-input";
 @import "../components/BFormSpinButton/spinbutton";
+@import "../components/BOffcanvas/offcanvas";
 
 .container,
 .container-fluid {
